refactor : CalculatorItemQueue 수정, Queue 제거
