+++ Calculator/Calculator/Model/Formula.swift
@@ -1,0 +1,17 @@
+//
+//  Formula.swift
+//  Calculator
+//
+//  Created by 김유진 on 2022/09/13.
+//
+
+import Foundation
+
+struct Formula {
+    var operands: CalculatorItemQueue
+    var operators: CalculatorItemQueue
+    
+    func result() -> Double {
+        
+    }
+}
